requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1 
async-timeout==4.0.3 
attrs==23.1.0 
cached-property==1.5.2 
certifi==2023.7.22 
charset-normalizer==3.2.0 
frozenlist==1.4.0 
greenlet==2.0.2 
idna==3.4 
inflection==0.5.1 
multidict==6.0.4 
mypy-extensions==1.0.0 
psycopg2-binary==2.9.7 
pyTelegramBotAPI==4.13.0 
python-dotenv==1.0.0 
requests==2.31.0 
SQLAlchemy==2.0.20 
sqlalchemy-orm==1.2.10 
typing-inspect==0.9.0 
typing_extensions==4.7.1 
urllib3==2.0.4 
yarl==1.9.2
